requirements.txt
# MCP (Model Context Protocol) Example Dependencies
anthropic>=0.7.0
python-dotenv>=1.0.0
requests>=2.31.0 